feat(PAT-80): Show info states and refactor to  modularize.
